% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hsa}
\alias{hsa}
\alias{hsa,mmu,rno}
\title{KdModelList for human miRNAs}
\value{
a \code{\link[scanMiR]{KdModelList}} list.
}
\description{
A list of binding affinity models (\code{\link[scanMiR]{KdModel}}) for human
('hsa'), mouse ('mmu'), or rat ('rno') miRbase 21.1 miRNAs. These represent
the miRNA 12-mer binding affinities predicted using the CNN from
McGeary, Lin et al. (2019).
}
