% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{KdModels-mmu}
\alias{KdModels-mmu}
\alias{KdModels-mmu,}
\alias{mmu}
\title{KdModelList for mouse miRNAs}
\value{
a \code{\link[scanMiR]{KdModelList}} list.
}
\description{
A list of binding affinity models (\code{\link[scanMiR]{KdModel}}) for mouse
miRbase 21.1 miRNAs. These represent the miRNA 12-mer binding affinities
predicted using the CNN from McGeary, Lin et al. (2019).
}
