% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKdModels.R
\name{getKdModels}
\alias{getKdModels}
\title{getKdModels}
\usage{
getKdModels(species = c("hsa", "mmu", "rno"))
}
\arguments{
\item{species}{Either "hsa" (human), "mmu" (mouse) or "rno" (rat)}
}
\value{
An object of class \code{\link[scanMiR]{KdModelList}}
}
\description{
Returns a collection of miRNA KdModels for the requested species.
}
\examples{
mods <- getKdModels("rno")
summary(mods)
}
